Corrigido lógica if

+++ src/service/EstoqueService.ts
@@ -40,13 +40,7 @@ export class EstoqueService {
   remover(codigo: number): boolean {
     const estoque = this.estoqueRepository.buscarCodigo(codigo);
 
-    if (!estoque) {
-      return false;
-    }
-
-    if (estoque.quantidade_emprestimo > 0) {
-      return false;
-    }
+    if (!estoque || estoque.quantidade_emprestimo > 0) return false;
 
     return this.estoqueRepository.remover(codigo);
   }
